requirements.txt
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
decorator==5.1.0
etherscan-python==2.1.0
idna==3.3
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycodestyle==2.8.0
Pygments==2.10.0
python-dotenv==0.19.1
requests==2.26.0
soupsieve==2.2.1
toml==0.10.2
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
